[model compression] reorganize quantizer files (#4341)
