+++ nni/algorithms/compression/pytorch/quantization/__init__.py
@@ -1,4 +1,12 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .quantizers import *
+from .bnn_quantizer import BNNQuantizer
+from .dorefa_quantizer import DoReFaQuantizer
+from .lsq_quantizer import LsqQuantizer
+from .native_quantizer import NaiveQuantizer
+from .observer_quantizer import ObserverQuantizer
+from .qat_quantizer import QAT_Quantizer
+
+
+__all__ = ['NaiveQuantizer', 'QAT_Quantizer', 'DoReFaQuantizer', 'BNNQuantizer', 'LsqQuantizer', 'ObserverQuantizer']
